Create user with role. Rearrange security classes. For future reference

+++ src/main/java/com/mentzikof/myPetClinic/repositories/PetClinicUserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.mentzikof.myPetClinic.model.PetClinicUser;
 
-public interface PetClinicUserRepository extends CrudRepository<PetClinicUser, Long> {
+public interface PetClinicUserRepository extends CrudRepository<PetClinicUser, Integer> {
 
 	PetClinicUser findByUsername(String username);
 }
